extract mapInterface component
