+++ jurassic-map-react/src/App.css
@@ -83,80 +83,6 @@ html {
   padding-left: 1px;
 }
 
-.mapTableMap {
-  position: relative;
-  grid-column: 2/3;
-  grid-row: 1/2;
-  min-height: 0;
-  max-height: 100%;
-}
-
-.react-transform-wrapper {
-  border-radius: 5px;
-}
-
-.map-table Canvas {
-  border-radius: 5px;
-  max-width: 100%;
-  max-height: 100%;
-  background: url('jpMap_2261x2492.png');
-  background-size: contain;
-}
-
-.tourVehicleInterface {
-  display: flex;
-  flex-flow: column nowrap;
-  z-index: 2;
-  position: absolute;
-  bottom: 2%;
-  left: 1%;
-  -webkit-font-smoothing: antialiased;
-  width: 15%;
-}
-
-.tourVehicleButton {
-  display: flex;
-}
-
-.tourVehicleButton img {
-  flex: 0 1 auto;
-  min-width: 35px;
-  padding: .75em;
-  cursor: pointer;
-}
-
-.tourVehicleButton p {
-  font-size: .85em;
-  text-transform: uppercase;
-  font-weight: 600;
-  margin: auto;
-}
-
-.mapZoomControls {
-  display: flex;
-  z-index: 2;
-  position: absolute;
-  bottom: 0px;
-  left: 54%;
-  transform: translate(-50%, -50%);
-  -ms-transform: translate(-50%, -50%);
-  -webkit-font-smoothing: antialiased;
-}
-
-.mapZoomControlBtn {
-  font-size: .85em;
-  text-transform: uppercase;
-  font-weight: 600;
-  cursor: pointer;
-  padding: .35em .55em;
-  background: white;
-  border: 1px solid grey;
-  border-radius: 5px;
-  margin-right: 5px;
-  margin-left: 5px;
-  -webkit-font-smoothing: antialiased;
-}
-
 .events-table {
   flex-direction: row;
   flex: 1 0 40vw;
